requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
asyncpg==0.29.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sqlalchemy==2.0.29
    # via
    #   -r requirements.in
    #   alembic
typing-extensions==4.11.0
    # via
    #   alembic
    #   sqlalchemy
